+++ tests/Cht/Augment/UsingLlm/AugmentTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace tests\Ssc\Btlr\Cht\Augment\UsingLlm;
 
 use Ssc\Btlr\Cht\Augment\UsingLlm\Augment;
+use Ssc\Btlr\Cht\Augment\UsingLlm\Augment\GetLastMessages;
 use Ssc\Btlr\Cht\Augment\UsingLlm\Log;
 use Ssc\Btlr\Cht\Augment\UsingLlm\Log\Source;
 use Ssc\Btlr\Framework\Filesystem\ReadFile;
@@ -20,29 +21,30 @@ class AugmentTest extends BtlrServiceTestCase
     {
         // Fixtures
         $userPrompt = 'Write code for me, please';
-        $augmentedPromptTemplateFilename = './var/cht/prompt_templates/augmented.txt';
+        $augmentedPromptTemplateFilename = './templates/cht/prompts/augmented.txt';
+        $lastMessagesFilename = './var/cht/logs/last_messages';
         $withConfig = [
             'augmented_prompt_template_filename' => $augmentedPromptTemplateFilename,
             'llm_engine' => 'chatgpt-gpt-3.5-turbo',
-            'logs_filename' => './var/cht/logs',
-            'log_filename_templates' => [
-                Source::USER_PROMPT => '%logs_filename%/conversation/%time%_000_%id%_%source%.json',
-                Source::AUGMENTED_PROMPT => '"%logs_filename%/augmentations/%time%_%id%.json"',
-                Source::MODEL_COMPLETION => '%logs_filename%/conversation/%time%_900_%id%_%source%.json',
-            ],
+            'last_messages_filename' => $lastMessagesFilename,
         ];
-        $augmentedPromptTemplate = 'Augmented %user_prompt%';
+        $lastMessages = "USER: Do you read me?\nBTLR: Affirmative dev, I read you\n";
+        $augmentedPromptTemplate = "%last_messages%USER: %user_prompt%\n";
         $thoseParameters = [
+            'last_messages' => $lastMessages,
             'user_prompt' => $userPrompt,
         ];
-        $augmentedPrompt = "Augmented {$userPrompt}";
+        $augmentedPrompt = "{$lastMessages}USER: {$userPrompt}\n";
 
         // Dummies
+        $getLastMessages = $this->prophesize(GetLastMessages::class);
         $log = $this->prophesize(Log::class);
         $readFile = $this->prophesize(ReadFile::class);
         $replace = $this->prophesize(Replace::class);
 
         // Stubs & Mocks
+        $getLastMessages->from($lastMessagesFilename)
+            ->willReturn($lastMessages);
         $readFile->in($augmentedPromptTemplateFilename)
             ->willReturn($augmentedPromptTemplate);
         $replace->in($augmentedPromptTemplate, $thoseParameters)
@@ -52,6 +54,7 @@ class AugmentTest extends BtlrServiceTestCase
 
         // Assertion
         $augment = new Augment(
+            $getLastMessages->reveal(),
             $log->reveal(),
             $readFile->reveal(),
             $replace->reveal(),